+++ tamagacha/View/BathroomView.swift
@@ -7,3 +7,15 @@
 
 import SwiftUI
 
+struct BathroomView: View {
+    @State var activeView: currentView
+    
+    var body: some View {
+        GeometryReader { geometry in
+            Text("Bathroom")
+                .frame(width: screenSize.size.width, height: screenSize.size.height, alignment: .center)
+        }
+        .background(Color.purple)
+        .edgesIgnoringSafeArea(.all)
+    }
+}
