+++ tamagacha/View/DeveloperToolsView.swift
@@ -7,3 +7,15 @@
 
 import SwiftUI
 
+struct DeveloperToolsView: View {
+    @Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
+    
+    var body: some View {
+        VStack {
+            Text("DevTools")
+            Button("Back") {
+                presentationMode.wrappedValue.dismiss()
+            }
+        }
+    }
+}
