+++ tamagacha/View/HomeView.swift
@@ -42,7 +42,7 @@ struct HomeView: View {
     var body: some View {
         VStack {
             ZStack {
-                HouseView(activeView: self.activeView)
+                RoomView(activeView: self.activeView)
                 StoreView(activeView: self.$activeView, navigateToSettings: $navigateToSettings, navigateToToyChest: $navigateToToychest, navigateToKitchen: $navigateToKitchen)
                     .offset(y: self.activeView == currentView.bottom ? 0 : screenHeight * 0.85)
                 BathroomView(activeView: self.activeView)
@@ -195,266 +195,3 @@ struct HomeView_Previews: PreviewProvider {
         HomeView()
     }
 }
-
-
-// MARK: ADD FILES
-
-// View Files
-/// HouseView
-struct HouseView: View {
-    @State var activeView: currentView
-    @State var isExpanded = false
-    
-    var scene: SKScene {
-        let scene = GameScene()
-        scene.size = CGSize(width: 400, height: 700)
-        scene.scaleMode = .fill
-        scene.anchorPoint = CGPoint(x: 0.5, y: 0.5)
-        return scene
-    }
-    
-    var body: some View {
-        GeometryReader { geometry in
-            ZStack {
-                VStack {
-                    SpriteView(scene: scene)
-                        .frame(width: 400, height: 700)
-                }
-                .frame(width: geometry.size.width, height: geometry.size.height, alignment: .center)
-                statView
-                swipeLeft
-            }
-        }
-        .background(Color.white)
-        .edgesIgnoringSafeArea(.all)
-    }
-    
-    var swipeLeft: some View {
-        ZStack {
-        Rectangle()
-            Image(systemName: "arrow.right")
-                .foregroundColor(Color.black)
-        }
-        .foregroundColor(.purple)
-        .cornerRadius(20, corners: [.topRight, .bottomRight])
-        .frame(width: 20, height: 100)
-        .padding(.trailing, screenWidth - 20)
-    }
-    
-    var statView: some View {
-        Group {
-            if isExpanded {
-                expandedStatView
-            } else {
-                collapsedStatView
-                    .padding(.leading, screenWidth - 50)
-            }
-        }
-        .onTapGesture {
-            withAnimation {
-                isExpanded.toggle()
-            }
-        }
-    }
-    
-    var expandedStatView: some View {
-        StatView()
-            .padding(.top, 70)
-            .padding(.bottom, 550)
-    }
-    
-    var collapsedStatView: some View {
-        ZStack {
-        Rectangle()
-            Text("Stats")
-                .foregroundColor(Color.red)
-        }
-        .cornerRadius(20, corners: [.topLeft, .bottomLeft])
-        .frame(width: 50, height: 100)
-//        .background(Color.blue)
-    }
-    
-    
-}
-
-/// StoreView
-struct StoreView: View {
-    @Binding var activeView: currentView
-    @Binding var navigateToSettings: Bool
-    @Binding var navigateToToyChest: Bool
-    @Binding var navigateToKitchen: Bool
-
-    var body: some View {
-        GeometryReader { geometry in
-            VStack {
-                ZStack {
-
-                    RoundedRectangle(cornerRadius: 20)
-                        .foregroundColor(.red)
-                    Text("Store")
-                    VStack {
-                        activeView == .center ? Image(systemName: "arrow.up") : Image(systemName: "arrow.down")
-                        if activeView == .center {
-                            HStack {
-
-                                Button("Kitchen") {
-                                    navigateToKitchen = true
-                                }
-                                Button("Toychest") {
-                                    navigateToToyChest = true
-                                }
-                                Button("Settings") {
-                                    navigateToSettings = true
-                                }
-
-                            }}
-                        
-                    }
-                    .padding(.bottom, 700)
-                }
-            }
-            .frame(width: geometry.size.width, height: geometry.size.height, alignment: .center)
-        }
-        .edgesIgnoringSafeArea(.all)
-        
-    }
-}
-
-/// BathroomView
-struct BathroomView: View {
-    @State var activeView: currentView
-    
-    var body: some View {
-        GeometryReader { geometry in
-            Text("Bathroom")
-                .frame(width: screenSize.size.width, height: screenSize.size.height, alignment: .center)
-        }
-        .background(Color.purple)
-        .edgesIgnoringSafeArea(.all)
-    }
-}
-
-/// SettingsView
-struct SettingsView: View {
-    @State var enterCode = ""
-    @State var navigateToDevTools = false
-    var devCode = "cheesepuffs"
-    
-    @Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
-    
-    var body: some View {
-        VStack {
-            Text("Settings")
-            Button("Back") {
-                presentationMode.wrappedValue.dismiss()
-            }
-            
-            TextField("Code", text: $enterCode)
-                .multilineTextAlignment(.center)
-                .autocorrectionDisabled()
-                .onSubmit {
-                    if enterCode.lowercased() == devCode {
-                        navigateToDevTools = true
-                    }
-                }
-            
-        }
-        .navigate(to: DeveloperToolsView(), when: $navigateToDevTools)
-
-    }
-    
-}
-
-/// ToychestView
-struct ToychestView: View {
-    
-    @Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
-    
-    var body: some View {
-        VStack {
-            Text("Toychest")
-            Button("Back") {
-                presentationMode.wrappedValue.dismiss()
-            }
-            
-        }
-    }
-    
-}
-
-/// KitchenView
-struct KitchenView: View {
-    
-    @Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
-    
-    var body: some View {
-        VStack {
-            Text("Kitchen")
-            Button("Back") {
-                presentationMode.wrappedValue.dismiss()
-            }
-        }
-    }
-    
-}
-
-/// DeveloperToolsView
-struct DeveloperToolsView: View {
-    @Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
-    
-    var body: some View {
-        VStack {
-            Text("DevTools")
-            Button("Back") {
-                presentationMode.wrappedValue.dismiss()
-            }
-        }
-    }
-}
-
-// Extentions
-/// RoundedRectangle
-extension View {
-    func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
-        clipShape( RoundedCorner(radius: radius, corners: corners) )
-    }
-}
-
-struct RoundedCorner: Shape {
-
-    var radius: CGFloat = .infinity
-    var corners: UIRectCorner = .allCorners
-
-    func path(in rect: CGRect) -> Path {
-        let path = UIBezierPath(roundedRect: rect, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
-        return Path(path.cgPath)
-    }
-}
-
-/// NavigateIfTrue
-extension View {
-    /// Navigate to a new view.
-    /// - Parameters:
-    ///   - view: View to navigate to.
-    ///   - binding: Only navigates when this condition is `true`.
-    func navigate<NewView: View>(to view: NewView, when binding: Binding<Bool>) -> some View {
-        NavigationView {
-            ZStack {
-                self
-                    .navigationBarTitle("")
-                    .navigationBarHidden(true)
-
-                NavigationLink(
-                    destination: view
-                        .navigationBarTitle("")
-                        .navigationBarHidden(true),
-                    isActive: binding
-                ) {
-                    EmptyView()
-                }
-            }
-        }
-        .navigationViewStyle(.stack)
-    }
-}
-
